[webdav] emit ERR_UNSUPPORTED_ACTION when performing fileSystemFreeSpace on a file
Performing a PROPFIND on a file (not a directory) results in a response containing empty text for the quota-used-bytes and quota-available-bytes properties (tested on wsgidav and Nextcloud).

These are then converted to 0 bytes.

To the caller this looks like the job succeeded and no space is available, but that's not actually the case.

To avoid this check whether the text is empty and don't process it then.

This will make the job emit ERR_UNSUPPORTED_ACTION and callers like CopyJob know to disregard it