[kprocessrunner] Also request an XDG activation token with StartupNotify=false
The semantics behind StartupNotify were defined for X11, it is unclear how they should translate to Wayland.

Currently when it is set to false we don't request an activation token. KWin suppresses visual startup feedback
for applications with StartupNotify=false.

There are cases where we do want an activation token but don't want the visual feedback. By requesting the token
(and leaving the question of visual feedback to the compositor) we can allow this use case